+++ src/main/java/com/github/hcsp/polymorphism/NameCollector.java
@@ -1,18 +1,18 @@
-package com.github.hcsp.polymorphism;
-
-import java.util.ArrayList;
-import java.util.List;
-import java.util.function.Consumer;
-
-public class NameCollector implements Consumer<User> {
-    private final List<String> names = new ArrayList<>();
-
-    @Override
-    public void accept(User user) {
-        names.add(user.getName());
-    }
-
-    public List<String> getNames() {
-        return names;
-    }
-}
+//package com.github.hcsp.polymorphism;
+//
+//import java.util.ArrayList;
+//import java.util.List;
+//import java.util.function.Consumer;
+//
+//public class NameCollector implements Consumer<User> {
+//    private final List<String> names = new ArrayList<>();
+//
+//    @Override
+//    public void accept(User user) {
+//        names.add(user.getName());
+//    }
+//
+//    public List<String> getNames() {
+//        return names;
+//    }
+//}
